feat(core): new all docs header (#12182)
