+++ packages/frontend/core/src/desktop/pages/workspace/all-collection/header.tsx
@@ -1,6 +1,6 @@
 import { IconButton } from '@affine/component';
+import { ExplorerNavigation } from '@affine/core/components/explorer/header/navigation';
 import { Header } from '@affine/core/components/pure/header';
-import { WorkspaceModeFilterTab } from '@affine/core/components/pure/workspace-mode-filter-tab';
 import { PlusIcon } from '@blocksuite/icons/rc';
 import clsx from 'clsx';
 
@@ -26,7 +26,7 @@ export const AllCollectionHeader = ({
           )}
         />
       }
-      center={<WorkspaceModeFilterTab activeFilter={'collections'} />}
+      left={<ExplorerNavigation active={'collections'} />}
     />
   );
 };
