+++ packages/frontend/core/src/desktop/pages/workspace/all-page-old/all-page.tsx
@@ -1,0 +1,96 @@
+import { useBlockSuiteDocMeta } from '@affine/core/components/hooks/use-block-suite-page-meta';
+import {
+  PageListHeader,
+  useFilteredPageMetas,
+  VirtualizedPageList,
+} from '@affine/core/components/page-list';
+import { GlobalContextService } from '@affine/core/modules/global-context';
+import { IntegrationService } from '@affine/core/modules/integration';
+import { WorkspacePermissionService } from '@affine/core/modules/permissions';
+import { WorkspaceService } from '@affine/core/modules/workspace';
+import type { Filter } from '@affine/env/filter';
+import { useI18n } from '@affine/i18n';
+import { useLiveData, useService } from '@toeverything/infra';
+import { useEffect, useState } from 'react';
+
+import {
+  useIsActiveView,
+  ViewBody,
+  ViewHeader,
+  ViewIcon,
+  ViewTitle,
+} from '../../../../modules/workbench';
+import { AllDocSidebarTabs } from '../layouts/all-doc-sidebar-tabs';
+import { EmptyPageList } from '../page-list-empty';
+import * as styles from './all-page.css';
+import { FilterContainer } from './all-page-filter';
+import { AllPageHeader } from './all-page-header';
+
+export const AllPage = () => {
+  const currentWorkspace = useService(WorkspaceService).workspace;
+  const globalContext = useService(GlobalContextService).globalContext;
+  const permissionService = useService(WorkspacePermissionService);
+  const integrationService = useService(IntegrationService);
+  const pageMetas = useBlockSuiteDocMeta(currentWorkspace.docCollection);
+  const [hideHeaderCreateNew, setHideHeaderCreateNew] = useState(true);
+  const isAdmin = useLiveData(permissionService.permission.isAdmin$);
+  const isOwner = useLiveData(permissionService.permission.isOwner$);
+  const importing = useLiveData(integrationService.importing$);
+
+  const [filters, setFilters] = useState<Filter[]>([]);
+  const filteredPageMetas = useFilteredPageMetas(pageMetas, {
+    filters: filters,
+  });
+
+  const isActiveView = useIsActiveView();
+
+  useEffect(() => {
+    if (isActiveView) {
+      globalContext.isAllDocs.set(true);
+
+      return () => {
+        globalContext.isAllDocs.set(false);
+      };
+    }
+    return;
+  }, [globalContext, isActiveView]);
+
+  const t = useI18n();
+
+  if (importing) {
+    return null;
+  }
+
+  return (
+    <>
+      <ViewTitle title={t['All pages']()} />
+      <ViewIcon icon="allDocs" />
+      <ViewHeader>
+        <AllPageHeader
+          showCreateNew={!hideHeaderCreateNew}
+          filters={filters}
+          onChangeFilters={setFilters}
+        />
+      </ViewHeader>
+      <ViewBody>
+        <div className={styles.body}>
+          <FilterContainer filters={filters} onChangeFilters={setFilters} />
+          {filteredPageMetas.length > 0 ? (
+            <VirtualizedPageList
+              disableMultiDelete={!isAdmin && !isOwner}
+              setHideHeaderCreateNewPage={setHideHeaderCreateNew}
+              filters={filters}
+            />
+          ) : (
+            <EmptyPageList type="all" heading={<PageListHeader />} />
+          )}
+        </div>
+      </ViewBody>
+      <AllDocSidebarTabs />
+    </>
+  );
+};
+
+export const Component = () => {
+  return <AllPage />;
+};
