+++ packages/frontend/core/src/desktop/pages/workspace/all-page/all-page-header.css.ts
@@ -1,0 +1,35 @@
+import { cssVarV2 } from '@toeverything/theme/v2';
+import { style } from '@vanilla-extract/css';
+
+export const header = style({
+  width: '100%',
+  height: '100%',
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+});
+
+export const actions = style({
+  display: 'flex',
+  alignItems: 'center',
+  gap: 16,
+});
+
+export const viewToggle = style({
+  backgroundColor: 'transparent',
+});
+export const viewToggleItem = style({
+  padding: 0,
+  fontSize: 16,
+  width: 24,
+  color: cssVarV2.icon.primary,
+  selectors: {
+    '&[data-state=checked]': {
+      color: cssVarV2.icon.primary,
+    },
+  },
+});
+export const viewToggleIndicator = style({
+  backgroundColor: cssVarV2.layer.background.hoverOverlay,
+  boxShadow: 'none',
+});
