+++ packages/frontend/core/src/desktop/pages/workspace/all-page/all-page-header.tsx
@@ -1,105 +1,75 @@
-import { usePageHelper } from '@affine/core/blocksuite/block-suite-page-list/utils';
-import {
-  AllPageListOperationsMenu,
-  PageDisplayMenu,
-  PageListNewPageButton,
-} from '@affine/core/components/page-list';
-import { Header } from '@affine/core/components/pure/header';
-import { WorkspaceModeFilterTab } from '@affine/core/components/pure/workspace-mode-filter-tab';
-import { WorkspaceDialogService } from '@affine/core/modules/dialogs';
-import { WorkbenchService } from '@affine/core/modules/workbench';
-import { WorkspaceService } from '@affine/core/modules/workspace';
-import { inferOpenMode } from '@affine/core/utils';
-import type { Filter } from '@affine/env/filter';
-import { track } from '@affine/track';
-import { PlusIcon } from '@blocksuite/icons/rc';
-import { useServices } from '@toeverything/infra';
-import clsx from 'clsx';
-import { useCallback } from 'react';
+import { type MenuProps, RadioGroup, type RadioItem } from '@affine/component';
+import { DocExplorerContext } from '@affine/core/components/explorer/context';
+import { ExplorerDisplayMenuButton } from '@affine/core/components/explorer/display-menu';
+import { DocListViewIcon } from '@affine/core/components/explorer/docs-view/doc-list-item';
+import { ExplorerNavigation } from '@affine/core/components/explorer/header/navigation';
+import type { ExplorerPreference } from '@affine/core/components/explorer/types';
+import { type Dispatch, type SetStateAction, useContext } from 'react';
 
-import * as styles from './all-page.css';
+import * as styles from './all-page-header.css';
 
-export const AllPageHeader = ({
-  showCreateNew,
-  filters,
-  onChangeFilters,
-}: {
-  showCreateNew: boolean;
-  filters: Filter[];
-  onChangeFilters: (filters: Filter[]) => void;
-}) => {
-  const { workspaceService, workspaceDialogService, workbenchService } =
-    useServices({
-      WorkspaceService,
-      WorkspaceDialogService,
-      WorkbenchService,
-    });
-  const workbench = workbenchService.workbench;
-  const workspace = workspaceService.workspace;
-  const { createEdgeless, createPage } = usePageHelper(workspace.docCollection);
+const views = [
+  {
+    label: <DocListViewIcon view="masonry" />,
+    value: 'masonry',
+    className: styles.viewToggleItem,
+  },
+  {
+    label: <DocListViewIcon view="grid" />,
+    value: 'grid',
+    className: styles.viewToggleItem,
+  },
+  {
+    label: <DocListViewIcon view="list" />,
+    value: 'list',
+    className: styles.viewToggleItem,
+  },
+] satisfies RadioItem[];
 
-  const handleOpenDocs = useCallback(
-    (result: {
-      docIds: string[];
-      entryId?: string;
-      isWorkspaceFile?: boolean;
-    }) => {
-      const { docIds, entryId, isWorkspaceFile } = result;
-      // If the imported file is a workspace file, open the entry page.
-      if (isWorkspaceFile && entryId) {
-        workbench.openDoc(entryId);
-      } else if (!docIds.length) {
-        return;
-      }
-      // Open all the docs when there are multiple docs imported.
-      if (docIds.length > 1) {
-        workbench.openAll();
-      } else {
-        // Otherwise, open the only doc.
-        workbench.openDoc(docIds[0]);
-      }
-    },
-    [workbench]
-  );
-
-  const onImportFile = useCallback(() => {
-    track.$.header.importModal.open();
-    workspaceDialogService.open('import', undefined, payload => {
-      if (!payload) {
-        return;
-      }
-      handleOpenDocs(payload);
-    });
-  }, [workspaceDialogService, handleOpenDocs]);
-
+const ViewToggle = () => {
+  const { view, setView } = useContext(DocExplorerContext);
   return (
-    <Header
-      left={
-        <AllPageListOperationsMenu
-          filterList={filters}
-          onChangeFilterList={onChangeFilters}
-          propertiesMeta={workspace.docCollection.meta.properties}
-        />
-      }
-      right={
-        <>
-          <PageListNewPageButton
-            size="small"
-            className={clsx(
-              styles.headerCreateNewButton,
-              !showCreateNew && styles.headerCreateNewButtonHidden
-            )}
-            onCreateEdgeless={e => createEdgeless({ at: inferOpenMode(e) })}
-            onCreatePage={e => createPage('page', { at: inferOpenMode(e) })}
-            onCreateDoc={e => createPage(undefined, { at: inferOpenMode(e) })}
-            onImportFile={onImportFile}
-          >
-            <PlusIcon />
-          </PageListNewPageButton>
-          <PageDisplayMenu />
-        </>
-      }
-      center={<WorkspaceModeFilterTab activeFilter={'docs'} />}
+    <RadioGroup
+      itemHeight={24}
+      gap={8}
+      padding={0}
+      items={views}
+      value={view}
+      onChange={setView}
+      className={styles.viewToggle}
+      borderRadius={4}
+      indicatorClassName={styles.viewToggleIndicator}
     />
+  );
+};
+
+const menuProps: Partial<MenuProps> = {
+  contentOptions: {
+    side: 'bottom',
+    align: 'end',
+    alignOffset: 0,
+    sideOffset: 8,
+  },
+};
+export const AllDocsHeader = ({
+  explorerPreference,
+  setExplorerPreference,
+}: {
+  explorerPreference: ExplorerPreference;
+  setExplorerPreference: Dispatch<SetStateAction<ExplorerPreference>>;
+}) => {
+  return (
+    <div className={styles.header}>
+      <ExplorerNavigation active="docs" />
+
+      <div className={styles.actions}>
+        <ViewToggle />
+        <ExplorerDisplayMenuButton
+          preference={explorerPreference}
+          onChange={setExplorerPreference}
+          menuProps={menuProps}
+        />
+      </div>
+    </div>
   );
 };
