+++ packages/frontend/core/src/desktop/pages/workspace/all-tag/header.tsx
@@ -1,6 +1,6 @@
+import { ExplorerNavigation } from '@affine/core/components/explorer/header/navigation';
 import { Header } from '@affine/core/components/pure/header';
-import { WorkspaceModeFilterTab } from '@affine/core/components/pure/workspace-mode-filter-tab';
 
 export const AllTagHeader = () => {
-  return <Header center={<WorkspaceModeFilterTab activeFilter={'tags'} />} />;
+  return <Header left={<ExplorerNavigation active={'tags'} />} />;
 };
