+++ packages/frontend/core/src/modules/feature-flag/constant.ts
@@ -283,6 +283,13 @@ export const AFFINE_FLAGS = {
     configurable: false,
     defaultState: isCanaryBuild,
   },
+  enable_new_all_docs_page: {
+    category: 'affine',
+    displayName: 'Enable New All Docs Page',
+    description: 'Use new all docs page',
+    configurable: isCanaryBuild,
+    defaultState: false,
+  },
 } satisfies { [key in string]: FlagInfo };
 
 // oxlint-disable-next-line no-redeclare
